Move SQL digest retriever into expression directory
