+++ executor/infoschema_reader.go
@@ -38,6 +38,7 @@ import (
 	"github.com/pingcap/tidb/domain"
 	"github.com/pingcap/tidb/domain/infosync"
 	"github.com/pingcap/tidb/errno"
+	"github.com/pingcap/tidb/expression"
 	"github.com/pingcap/tidb/infoschema"
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/meta/autoid"
@@ -2205,11 +2206,11 @@ func (e *tidbTrxTableRetriever) retrieve(ctx context.Context, sctx sessionctx.Co
 	var res [][]types.Datum
 	err = e.nextBatch(func(start, end int) error {
 		// Before getting rows, collect the SQL digests that needs to be retrieved first.
-		var sqlRetriever *SQLDigestTextRetriever
+		var sqlRetriever *expression.SQLDigestTextRetriever
 		for _, c := range e.columns {
 			if c.Name.O == txninfo.CurrentSQLDigestTextStr {
 				if sqlRetriever == nil {
-					sqlRetriever = NewSQLDigestTextRetriever()
+					sqlRetriever = expression.NewSQLDigestTextRetriever()
 				}
 
 				for i := start; i < end; i++ {
